requirements.txt
streamlit
opencv-python
ultralytics
Pillow
numpy<2.0.0 